Move the model from pyanaconda.storage
Move the class InstallerStorage and the related support to the Storage module.
