+++ pyanaconda/modules/storage/checker/utils.py
@@ -602,7 +602,7 @@ class StorageChecker(object):
         This function is called at the end of partitioning so that we can make
         sure you don't have anything silly (like no /, a really small /, etc).
 
-        :param storage: an instance of the :class:`pyanaconda.storage.InstallerStorage` class to check
+        :param storage: the storage object to check
         :param constraints: an dictionary of constraints that will be used by
                checks or None if we want to use the storage checker's constraints
         :param skip: a collection of checks we want to skip or None if we don't
