+++ pyanaconda/modules/storage/devicetree/model.py
@@ -33,10 +33,10 @@ from pyanaconda.core import util
 from pyanaconda.modules.storage.bootloader import BootLoaderFactory
 from pyanaconda.core.configuration.anaconda import conf
 from pyanaconda.core.constants import shortProductName
-from pyanaconda.storage.fsset import FSSet
+from pyanaconda.modules.storage.devicetree.fsset import FSSet
 from pyanaconda.modules.storage.devicetree.utils import download_escrow_certificate, \
     find_live_backing_device
-from pyanaconda.storage.root import find_existing_installations
+from pyanaconda.modules.storage.devicetree.root import find_existing_installations
 from pyanaconda.modules.common.constants.services import NETWORK
 
 import logging
