+++ pyanaconda/storage/initialization.py
@@ -26,7 +26,7 @@ from blivet.static_data import luks_data
 
 from pyanaconda.anaconda_logging import program_log_lock
 from pyanaconda.core.configuration.anaconda import conf
-from pyanaconda.storage.osinstall import InstallerStorage
+from pyanaconda.modules.storage.devicetree.model import InstallerStorage
 from pyanaconda.modules.storage.platform import platform
